requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.1
gspread==6.1.2
google-auth==2.29.0
google-generativeai==0.7.2
python-dateutil==2.9.0
